requirements.txt
flask-cors
flask
pymongo[srv]
python-dotenv
gunicorn
scikit-learn
openai